Corrects an error in dataset loading that also impacted the published paper's results.
The nested loop runs through all years (outer loop) and through all fires in these years (inner loop). If the inner loop found the searched-after fire, it would break. However, the outer loop would continue to the next year and update found_fire_year and found_fire_name before breaking in that second year.

Based on initial experiments, the corrected dataloader leads to slightly higher performance, but the trends in the results are basically the same as those reported in the paper.